Fix calculation of maximum time value.

+++ src/posix/luaosi/timedef.h
@@ -4,7 +4,7 @@
 
 typedef void losi_TimeDriver;
 
-#define LOSI_TIME_MAX (INT_MAX + INT_MAX / 1000000000)
+#define LOSI_TIME_MAX (INT_MAX + (double)INT_MAX / (double)1000000000)
 
 #define LOSI_DISABLE_TIMEDRV
